open/close refactoring and less wordy status commands
